split griddle logic
