+++ src/gribbles/logic/gameInit.js
@@ -1,7 +1,6 @@
-import { isKnown } from "../../common/isKnown";
 import { letterPool } from "../../common/letterPool";
-import { getSurroundingIndexes } from "./updateGameState";
 import { shuffleArray } from "../../common/shuffleArray";
+import { findAllWords } from "./findAllWords";
 
 function getLetters(gridSize) {
   // Given the distribution of letters in the word list
@@ -12,51 +11,6 @@ function getLetters(gridSize) {
   return chosenLetters;
 }
 
-function findAllWords({ grid, minWordLength, easyMode }) {
-  let foundWords = [];
-
-  const neighborIndexes = grid.map((_, index) =>
-    getSurroundingIndexes({ index: index, gridSize: Math.sqrt(grid.length) })
-  );
-
-  function checkSurrounding(currentIndex, currentWord, visitedIndexes) {
-    let surroundingIndexes = neighborIndexes[currentIndex];
-    for (let surroundingIndex of surroundingIndexes) {
-      // if the index has already been used, skip
-      if (visitedIndexes.includes(surroundingIndex)) {
-        continue;
-      }
-      const newWord = currentWord + grid[surroundingIndex];
-      const { isPartial, isWord, isEasy } = isKnown(newWord);
-
-      if (easyMode) {
-        if (isEasy && newWord.length >= minWordLength) {
-          foundWords.push(newWord);
-        }
-      } else {
-        if (isWord && newWord.length >= minWordLength) {
-          foundWords.push(newWord);
-        }
-      }
-      if (isPartial) {
-        checkSurrounding(
-          surroundingIndex,
-          newWord,
-          visitedIndexes.concat(surroundingIndex)
-        );
-      }
-    }
-  }
-
-  for (let startingIndex = 0; startingIndex < grid.length; startingIndex++) {
-    checkSurrounding(startingIndex, grid[startingIndex], [startingIndex]);
-  }
-
-  const uniqueFoundWords = new Set(foundWords);
-
-  return Array.from(uniqueFoundWords).sort();
-}
-
 function getPlayableLetters({ gridSize, minWordLength, easyMode }) {
   // Select letters and make sure that the computer can find at least
   // 50 words (standard mode) or 20 words (easy mode)
@@ -65,7 +19,7 @@ function getPlayableLetters({ gridSize, minWordLength, easyMode }) {
   let foundPlayableLetters = false;
   let letters;
   let allWords;
-  while (!foundPlayableLetters) {
+  while (!foundPlayableLetters) {//todo how to prevent infinite loop?
     letters = getLetters(gridSize);
     allWords = findAllWords({
       grid: letters,
@@ -80,7 +34,7 @@ function getPlayableLetters({ gridSize, minWordLength, easyMode }) {
   return [letters, allWords];
 }
 
-export function getInitialGameState({ gridSize, minWordLength, easyMode }) {
+export function gameInit({ gridSize, minWordLength, easyMode }) {
   // use the specified settings, otherwise check local storage, otherwise use default
   gridSize = gridSize || JSON.parse(localStorage.getItem("gridSize")) || 4;
   minWordLength =
